Promote QEMU actions to shared package and fix colon suffix in ActionCharDevices
Move 8 duplicate QEMU actions from completer-specific action/ packages to
pkg/actions/tools/qemu/ so they are available as macros (tools.qemu.*).
Remove hardcoded .Suffix(":") from ActionCharDevices per the guideline that
suffix/prefix modifiers belong at the call site, not in action definitions.
